Only style changes
No changes in the program, only in the style

+++ frequency.h
@@ -3,32 +3,32 @@
 
 #include<iostream>
 #include<cmath>
+#include<string>
 
 using namespace std;
 
 int Frequency(int i=0) {
-    double f;
-    if (i==0){
+    int f;
+    if (i == 0){
     cout << "Hello," << endl;
-    cout
-            << "Considering your repayment schedule, how often would you like install repayment rates install payments per year?\n";
+    cout << "Considering your repayment schedule, how often would you like install repayment rates install payments per year?\n";
     cout << "Available rate options are:\n";
     }
     cout << " 1 = annually \n 2 = semi-annually \n 3 = quarterly \n 4 = monthly\n";
     cout << "Insert your choice by the provided numbers: \n";
     cin >> f;
 
-    if (f==1){
+    if (f == 1) {
     cout<<"You selected annual repayment rate\n";
-    }else if(f==2){
+    } else if(f == 2) {
         cout<<"You selected semi-annual repayment rates\n";
-    }else if(f==3){
+    } else if(f == 3) {
         cout<<"You selected quarterly repayment rates\n";
-    }else if(f==4){
+    } else if(f == 4) {
         cout<<"You selected monthly repayment rates\n";
-    }else {
+    } else {
         cout<<"Please insert a valid input.\n\n";
-        Frequency(i=1);
+        Frequency(i = 1);
     }
     
     return f;
